requirements.txt
# Core dependencies
llama_parse>=0.1.0
llama_index>=0.10.0
python-dotenv>=1.0.0
nest_asyncio>=1.5.6
streamlit>=1.28.0
google-generativeai>=0.3.0

# Alternative PDF parsers
PyPDF2>=3.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Microsoft MarkItDown
markitdown[all]>=0.1.0 